+++ src/store/actions/navigation.js
@@ -1,0 +1,8 @@
+import {SET_PATH} from "./types";
+
+export const setPath = dispatch => (data) => {
+  dispatch({
+    type: SET_PATH,
+    payload: data
+  })
+}
